# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mkutano Convening Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      parameters:
        - name: conveningId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConveningDocument"
    post:
      operationId: createDocument
      tags: [documents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConveningDocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConveningDocument"
  /documents/{id}:
    patch:
      operationId: updateDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: conveningId
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConveningDocumentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConveningDocument"
    delete:
      operationId: deleteDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: conveningId
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /templates:
    get:
      operationId: listTemplates
      tags: [documents]
      parameters:
        - name: conveningId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OutreachTemplate"
    post:
      operationId: createTemplate
      tags: [documents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutreachTemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachTemplate"
  /templates/{id}:
    patch:
      operationId: updateTemplate
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: conveningId
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutreachTemplateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachTemplate"
    delete:
      operationId: deleteTemplate
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: conveningId
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Object Storage ───────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    ConveningDocument:
      type: object
      required: [id, type, title, url, version, createdAt]
      properties:
        id:
          type: string
        conveningId:
          type: [string, 'null']
        type:
          type: string
          enum: [Deck, GuidingDoc, Prospectus, Report, AideMemoire, Communique, Contract, Other]
        title:
          type: string
        url:
          type: string
        version:
          type: integer
        notes:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    ConveningDocumentInput:
      type: object
      required: [conveningId, type, title, url]
      properties:
        conveningId:
          type: string
        type:
          type: string
          enum: [Deck, GuidingDoc, Prospectus, Report, AideMemoire, Communique, Contract, Other]
        title:
          type: string
        url:
          type: string
        version:
          type: integer
        notes:
          type: string
    OutreachTemplate:
      type: object
      required: [id, name, category, scope, bodyMarkdown, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [PartnerLetter, SponsorLetter, PensionFundLetter, AdvisoryBoardConceptNote, AdvisoryBoardInvitation, SteeringCommitteeConceptNote, Other]
        scope:
          type: string
          enum: [Global, Convening]
        conveningId:
          type: [string, 'null']
        bodyMarkdown:
          type: string
        createdAt:
          type: string
          format: date-time
    OutreachTemplateInput:
      type: object
      required: [conveningId, name, category, bodyMarkdown]
      properties:
        name:
          type: string
        category:
          type: string
          enum: [PartnerLetter, SponsorLetter, PensionFundLetter, AdvisoryBoardConceptNote, AdvisoryBoardInvitation, SteeringCommitteeConceptNote, Other]
        scope:
          type: string
          enum: [Global, Convening]
        conveningId:
          type: string
        bodyMarkdown:
          type: string
    ConveningDocumentUpdate:
      type: object
      properties:
        title:
          type: string
        notes:
          type: string
        version:
          type: integer
        url:
          type: string
...